refactor: rename titleitem directory to automatictitleitem
