+++ packages/admin/src/common/toolbar/automatictitleitem/ToolbarAutomaticTitleItem.styles.tsx
