+++ packages/admin/src/common/toolbar/automatictitleitem/ToolbarAutomaticTitleItem.tsx
