+++ packages/admin/src/common/toolbar/index.ts
@@ -2,8 +2,8 @@ export * from "./Toolbar";
 export { CometAdminToolbarClassKeys, ToolbarThemeProps } from "./Toolbar.styles";
 export * from "./item/ToolbarItem";
 export { CometAdminToolbarItemClassKeys } from "./item/ToolbarItem.styles";
-export * from "./titleitem/ToolbarAutomaticTitleItem";
-export { ToolbarAutomaticTitleItemThemeProps } from "./titleitem/ToolbarAutomaticTitleItem.styles";
+export * from "./automatictitleitem/ToolbarAutomaticTitleItem";
+export { ToolbarAutomaticTitleItemThemeProps } from "./automatictitleitem/ToolbarAutomaticTitleItem.styles";
 export * from "./breadcrumb/ToolbarBreadcrumbs";
 export { ToolbarBreadcrumbsThemeProps } from "./breadcrumb/ToolbarBreadcrumbs.styles";
 export * from "./backbutton/ToolbarBackButton";
